Allow multiple directories of 'symlinked_files'
Symbolic links will be created for all files found within directories
named with the prefix 'symlinked_files'. These links will be given a dot
prefix as before.

The files previously in symlinked_files/ have been moved to one of
following new directories:

- The 'symlinked_files-tooling' directory contains the dotfiles used
  directly by programs such as git and vim.

- The 'symlinked_files-common' directory contains the dotfiles included
  by all shell startup files to set paths, aliases, environment
  variables, etc.

- The 'symlinked_files-bash' directory contains just the dotfiles used
  to configure Bash.

- The 'symlinked_files-zsh' directory contains just the dotfiles used
  to configure the Z shell.

This paves the way for adding support to the installer to fetch and link
machine or context specific repos which add additional dotfiles later.