Avoid adding duplicates to fpath
Usually Brew installs completions in a location which macOS adds to the
fpath by default. By searching the fpath array for that location and
returning the index of the match or array lenght plus one in the case of
a failure to find a match we only add it when necessary.

See https://zsh.sourceforge.io/Doc/Release/Parameters.html#Subscript-Flags.