Switch to project directory if run from elsewhere
The script only functions correctly when run from within the project
directory. To avoid erroneous behaviour it will now change the CWD to
the directory the project was cloned into if executed from elsewhere.

In particular it is important to ensure that it does not attempt to
create the dotfile symlinks to a path which itself is a symlink, which
the use of readlink avoids.