Clear git pager content on exit
When a git command invokes the pager the content will be cleared from
the terminal on exit, in the same way that git log is configured to. The
pager will only be used should the content fill more than a single
screen (unlike git log, which always uses the pager).